- (Diep) - Refactored some more out of the game-engine - Renamed some files and moved buttons files to a buttons folder.
