+++ src/app/diep/ui/buttons/diep.button-interaction.service.ts
@@ -1,14 +1,14 @@
 import { Injectable } from '@angular/core';
-import { DiepGameEngineService } from '../engine/diep.game-engine.service';
-import { DiepButton } from '../core/diep.interfaces';
-import { DiepQuadriviumMenu } from './main-menu/quadrivium/diep.quadrivium-menu';
-import { DiepAchievementMenu } from './main-menu/achievements/diep.achievement-menu';
-import { DiepAchievementNavigator } from './main-menu/achievements/diep.achievement-nav-bar';
-import { DiepMainMenu } from './main-menu/diep.main-menu';
-import { DiepPauseOverlay } from './overlays/pause-overlay';
-import { DiepGameOverOverlay } from './overlays/game-over-overlay';
-import { DiepHealthBarRenderer } from './hud/diep.health-bar-renderer';
-import { DiepUpgradeMenuRenderer } from './hud/upgrade-menu/diep.upgrade-menu-renderer';
+import { DiepGameEngineService } from '../../engine/diep.game-engine.service';
+import { DiepButton } from '../../core/diep.interfaces';
+import { DiepQuadriviumMenu } from '../main-menu/quadrivium/diep.quadrivium-menu';
+import { DiepAchievementMenu } from '../main-menu/achievements/diep.achievement-menu';
+import { DiepAchievementNavigator } from '../main-menu/achievements/diep.achievement-nav-bar';
+import { DiepMainMenu } from '../main-menu/diep.main-menu';
+import { DiepPauseOverlay } from '../overlays/pause-overlay';
+import { DiepGameOverOverlay } from '../overlays/game-over-overlay';
+import { DiepHealthBarRenderer } from '../hud/diep.health-bar-renderer';
+import { DiepUpgradeMenuRenderer } from '../hud/upgrade-menu/diep.upgrade-menu-renderer';
 
 @Injectable({ providedIn: 'root' })
 export class DiepInteractionService {
