+++ src/app/diep/ui/main-menu/achievements/diep.achievement-menu.ts
@@ -2,7 +2,7 @@ import { Achievement, DiepButton } from '../../../core/diep.interfaces';
 import { AchievementListSorter } from './achievement-list.sorter';
 import { AchievementCardRenderer } from './achievement-card.renderer';
 import { DiepAchievementNavigator } from './diep.achievement-nav-bar';
-import { DiepButtonRenderer } from '../../diep.button-renderer';
+import { DiepButtonRenderer } from '../../buttons/diep.button-renderer';
 
 export class DiepAchievementMenu {
   private static scrollY = 0;
